% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rb_images.R
\name{get_image_urls}
\alias{get_image_urls}
\alias{open_rb_images}
\alias{download_rb_images}
\title{Get specimen image URLs, open them in a browser or download and store them
locally.}
\usage{
get_image_urls(specimens = NULL, scientific_name = NULL, family = NULL,
  barcode = NULL, max = 5, random = FALSE, width = 600)

open_rb_images(specimens = NULL, scientific_name = NULL, family = NULL,
  barcode = NULL, max = 5, random = FALSE, width = 600)

download_rb_images(specimens = NULL, scientific_name = NULL,
  family = NULL, barcode = NULL, width = 3000, random = FALSE,
  max = 50)
}
\arguments{
\item{specimens}{A data frame returned by \code{\link{search_rb}}.}

\item{scientific_name}{A character vector of names to be searched.}

\item{family}{A character vector of families.}

\item{barcode}{A character vector of barcodes.}

\item{max}{Maximum number of results to display or download.}

\item{random}{Logical indicating if random images should be downloaded or open when \code{max} is
lower than the number of search results. The default is to display the first \code{max} results.}

\item{width}{The width of the downloaded or open images in pixels.}
}
\description{
These functions provide functionality to handle images from the RB herbarium. The
easiest way to use them is to get the object returned by \code{\link{search_rb}} and
pass it to these functions. Alternatively, you can search by scientific name, family, and barcode.
}
