% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{RB}
\alias{RB}
\alias{rb_env}
\alias{RB-package}
\alias{RB-package}
\title{Package RB}
\format{An object of class \code{environment} of length 0.}
\usage{
rb_env
}
\description{
Query the RB herbarium dataset and download specimens images
}
\details{
This packages contain a set of tools for querying and downloading images
from the RB herbarium dataset.
}
\keyword{datasets}
