% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rb_data.R
\name{rb_latest_data}
\alias{rb_latest_data}
\title{Download the Darwin Core Archive file from http://ipt.jbrj.gov.br/jbrj/}
\usage{
rb_latest_data()
}
\description{
These functions are used mainly internally to manage the DWC-A files from
the RB IPT repository. All files are downloaded to a local cache and will be
reused as long as the cache isn't purged.
}
