% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_rb.R
\name{search_rb}
\alias{search_rb}
\title{Search specimens in the RB collection.}
\usage{
search_rb(scientific_name = NULL, family = NULL, genus = NULL,
  county = NULL, barcode = NULL, state = NULL, with_images = FALSE,
  with_coordinates = FALSE, year = NULL)
}
\arguments{
\item{scientific_name}{A character vector with one or more scientific names.}

\item{family}{A character vector with one or more families.}

\item{genus}{A character vector with one or more genera.}

\item{county}{A character vector with one or more counties.}

\item{barcode}{A character vector with one or more catalog numbers.}

\item{state}{A character vector with one or more Brazilian states (full names only).}

\item{with_images}{A logical indicating whether to return only specimens with images.}

\item{with_coordinates}{A logical indicating whether to return only specimens with geographical
coordinates.}

\item{year}{A numerical or character vector indicating years to refine the search.}
}
\description{
This function filters a local copy of the RB collection according to
user-defined arguments. The returned object can them be used to visualize or download
register images.
}
\examples{
\dontrun{
search_rb("Miconia albicans")
search_rb(family = "Melastomataceae", county = "Itirapina")
search_rb(genus = "Myrcia", state = "Rio de Janeiro")
search_rb("Miconia rubiginosa", years = 2001:2003, with_images = TRUE)
}
}
