% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{RB}
\alias{RB}
\alias{RB-package}
\alias{RB-package}
\title{Package RB}
\description{
Query the RB herbarium dataset and download specimens images
}
\details{
This packages contain a set of tools for querying and downloading images
from the RB herbarium dataset.
}
